epan: apply endianness fixes
Historically a boolean (TRUE/FALSE) was being used for little/big
endian encoding designation. With the introduction of the encoding
enum, for backward compatibility reasons, the boolean interpretation
of this encoding designation is supported to determine endianness.
This leads to a lot of code using ENC_NA where ENC_BIG_ENDIAN is
intended, because these 'promote' to the same boolean interpretation.

With this commit a series of identified incorrect endianness
designations are corrected. A few other endianness related are issues
are addressed as well.