Compatibility header for using zlib and zlib-ng
Extract nearly-identical boilerplate code for detecting and using either
zlib-ng or zlib from 9 files and put it into its own header. Add a
comment explaining its purpose and motivation.

Add a typedef for `zlib_streamp` which is either `z_streamp` or
`zng_streamp` as appropriate. This is analogous to existing code for
`zlib_stream`. Remove extra `#ifdef`s in `packet-websocket.c`,
`packet-spdy.c`, and `wiretap/file-wrappers.c` for determining
which to use in favour of this new typedef.

Fix an apparent bug in `packet-blip.c` where the user preference for
maximum uncompressed body size was exposed if zlib was in use but
hidden if zlib-ng was in use.

In files which directly use other zlib functions, replace tests for
either `HAVE_ZLIB` or `HAVE_ZLIBNG` with a slightly briefer test for
`USE_ZLIB_OR_ZLIBNG`. (Inspired by !16047.)

Files which have this test but don't directly use zlib were left
untouched in order to reduce the diff size.