elastic-mapping: Add mandatory index_patterns name
ElasticSearch index templates require a index_pattern name that is an
array of strings:

https://www.elastic.co/guide/en/elasticsearch/reference/current/indices-templates-v1.html

Our current format is the v1 legacy "_template" API. The newer
"_index_template" API would also require that key, but also need
other changes to the structure.

Update by adding the name with the value the appropriate array
containing as wildcard the index names that we write in the ek (and
json) output from tshark.

Fix #19657 (combined with fe1217ac75d57b44a3b and 54ebf23913a7637631745)

There's still some other points raised in that issue, but they're
duplicates of other issues.