dumpcap: As a child process, never wait for a keypress to detach console
If dumpcap is running as a capture child, never wait for the user to
press a key before detaching the console.  Make sure this is false
after initializing the logging (whether it was set there, or the
console_wait static and the destroy_console called atexit() inherited
from the parent process).

Fix #20434