raw, ipv4, ipv6: handle raw IPv4 and IPv6 link-layer types like raw IP.
Handle "raw IPv4" and "raw IPv6" link-layer types similarly to the "raw
IP" link-layer type; the only difference is that "raw IPv4" means
"everything is IPv4" and "raw IPv6" means "everything is raw IPv6", as
opposed to "raw IP" meaning "this is raw IP, look at the version number
in thr network-layer header to determine whether it's IPv4 or IPv6".

This means that raw IPv4 and IPv6 will get a "raw link layer" header
similar to raw IP, and a new tvb will be handed to the IPv4 or IPv6
dissector so that, if the length is set by the IPv4 or IPv6 dissector,
it won't affect the top-layer tvb.

Addresses more of issue #20423.