ippusb: hook interface protocol instead of bulk type
The bulk handler is run after heuristic dissectors. This caused
AT dissector to eat some packets, and possibly the injection
of newline packets was added to workaround this.

Use the official IPPUSB protocol and HP vendor specific protocols
instead to get always called.

This allows to remove the newline synthesis that worked incorrectly
depending on how the HTTP traffic was split over USB bulk transfers.

Always have the dissector also report the packet taken as we know
it belongs to us. Simplify things.