ippusb: remove broken herustic based on assumed ipp operation id
The status_code is expected to be the IPP operation-type, but it
is not. While IPP operation-type is the bytes 2-3 of IPP packet,
the packet seen by IPPUSB dissector is not straight up the IPP
PDU. This is obvious as the status_code is used in same if block
with is_http_header() which assumes the packet starts with HTTP
headers.

Because the code breaks on small packets (reads beyond end of
packet) and reads from incorrect location anything based on it
is wrong. Remove the heuristic and any assumptions made based
on it.

This fixes decoding of dumps with small USB packets.