ippusb: Show what frame payload was reassembled in
We don't need to create a new MSP when adding to an existing one.
Just save the old running length on the first pass.

By doing this, we can call retrieve the fragment head and call
process_reassembled_data for fragments other than the last PDU,
and add the "Reassembled payload in frame:" FT_FRAMENUM to the tree.

Ping #20438