wiretap: Open JSON files requiring more tokens
jsmn_validate puts a maximum number of tokens allowed. Add a function
to call jsmn_parse without calling strlen (so it both works without a
null terminated string and is more efficient when we already know the
length) and call that instead.

Add a comment that we might want to change jsmn_validate to just set
a maximum length, instead of allocating memory for an array of tokens
that are filled out and then freed.

Even if jsmn_validate is changed, the new function would still be useful
for parsing partial JSON, whether truncated files or if the heuristics
decided to make a decision after less than an entire file.

Fix #17589