conversations_taps: Add preference for machine readable output
Add a preference for the conversations taps (both CLI and Wireshark)
for machine readable output. A preference probably makes a bit more
sense because people probably want to persist this. It's also specific
to one tap, so a tshark long option doesn't make as much sense, but
adding optional flags (as opposed to mandatory choices) becomes far
to difficult for the stat flags.

This preference is respected in the endpoints taps and Dialog as well.

Add a checkbox in the GUI to display the machine readable values,
set by default to the preference value.

Fix #18321