TLS: Display the tls-record-number
Save the TLS record number upon successful decryption and display it
as a generated field. Note that it would take greater changes to
save the record number attempted for failed decryption, and it would
not necessarily be accurate, especially for TLS 1.3 (because of the
record number resetting after switching from the handshake key to the
appdata key, but the Finished message indicating that the key should
switch is encrypted.)

We were already saving the (lower 16 bits of the) decrypted record
sequence number for DTLS 1.3, so simplify and use some common code,
storing the TLS and DTLS record sequence numbers the same way.

Fix #20446