dumpcap: Handle pcapng with different endianness for a single input pipe
If dumpcap is reading from a single pipe (extcap or not) supplying pcapng,
then the SHBs and IDBs are passed through directly. That means that
supporting the capture (whether all of one byte order different than
the host's, or mixed-endian) is up to libwiretap and libwireshark,
not dumpcap. All dumpcap has to do in that case is swap the block
type and block length after reading each block header in the section.
Implement that.

The multiple source case is harder because dumpcap generates its own SHB,
and thus has to rewrite the byte order of any elements of any block header
(including options) that have a byte order that corresponds to that given
in the SHB. Check for that case, and continue to produce an error. Swap
around the logic so we assume we're doing passthrough initially if there's
only one interface, so we don't fail unnecessarily. (Note we can get a SHB
of a different endianess on the same pipe later.)

Some link-layer interfaces, like Linux USB, or Linux SLL when carrying
CAN, have pseudoheaders that use host byte order. Those probably cause
problems in pcap as well; add a comment about that.

Add a test (note this test will timeout without
commit ac0ef1832810f9d79eac4b2d5eac1d88ca353495)

Ping #15754. Ping #15707.