dumpcap: Set pipe error status on non-initial SHB read failure
If a dumpcap is reading from a pipe, and reading a SHB subsequent
to the first one fails, set the pipe error status and return with
failure, instead of continually trying to read from the pipe.

This can happen, for example, if reading a pipe pcapng with mixed
BE/LE SHBs, where the first SHB is the same Endianness as the host,
but a later SHB is not.

Ping #15707


(cherry picked from commit ac0ef1832810f9d79eac4b2d5eac1d88ca353495)

Co-authored-by: John Thacker <johnthacker@gmail.com>