EDHOC: Don't use the wrong printf specifier
g_bytes_get_size() returns a gsize, which is almost always a size_t.
(Subject to a glib bug on some obscure platforms.)
There are platforms where a long is 32-bit and a gsize/size_t is
a 64 bit unsigned long long, and compilers that will warn if %ld is used
instead of %zu in that case.

Per the comment in !19291, just remove logging the size.