http3: remove static memory pointer
Remove static pointer from `get_header_field_pstr()`, and code which
used its target as a buffer.

The change in !19358 introduced a static pointer to memory allocated in
`wmem_file_scope()`. It held this pointer even after the target memory
had been freed due to closing or changing files, resulting in a dangling
pointer and causing memory errors.

The memory in question was being used as scratch space in order to
reduce per-packet memory reallocations, but `pinfo->pool` was
designed to gracefully and transparently handle that consideration.
Remove the static buffer and just use `pinfo->pool`.