dfilter: Have vals() return the same as column "display as strings"
Have the vals() display filter function / value string instruction
return the same value as what custom columns use for "display as
strings", i.e., call proto_item_fill_display_label()

This doesn't change anything at all in the cases where it's been
returning something from a value string; it only adds a numeric
string for the cases where there is no entry in the value string.

Using this to output the same thing as the custom columns do
"as strings" is the most common use case for this function (among
other things, it replaces the various _ws.col hacks), and it probably
should return something non-NULL for each field that is present.

This has the benefit of using common code and having Wireshark and
TShark produce the same output in different places.

Fix #20456