FTDI FT: Simplify reassembly
Simplify reassembly and avoid the use of composite tvbs.

Since the fragment IDs are unique (being based upon the first frame),
we don't need to remove completed fragments and put them in a reassembled
table. This allows the use of fragment_set_partial_reassembly, much like
the TCP dissector.

fragment_truncate has been introduced for TCP, while allows shortening
the last added fragment, addressing another concern (when a packet
contains both the end of a previous reassembly and the beginning of
a new one.)

By splitting process_reassembled_data into its component actions,
we can add the fragment tree on the first pass only after the
subdissector has reported success, avoiding the hack of changing the
current layer number.

We can avoid an O(N) calculation of the current total data length
by using it when we have it (which is every time except for the
first segment added.)

This is about 40% faster opening the ft232h-scan.pcapng file from
issue #11743, and avoids an issue with the fuzzed data in #20441
trying to construct a composite TVB with 3753 fragments, all of
which are themselves composite TVBs, and too much recursion.

Fix #20441