Qt: Make it possible to view JSON data as pretty-printed JSON
In the Qt UI, add BaseDataSourceView and JsonDataSourceView widgets.
JsonDataSourceView pretty-prints JSON and is primarily intended for
logs, but should be useful for JSON objects in packet data as well.

Make ByteViewText and JsonDataSourceView subclasses of
BaseDataSourceView. We should rename ByteViewText to ByteDataSourceView
at some point.

In ByteViewTab, create a ByteViewText or JsonDataSourceView based on the
data source media type.

In epan, add a media type attribute to data sources, and define octet
stream and JSON media types.

In the Falco Bridge dissector, set the log data media type to JSON when
dissecting CloudTrail logs.