http3: dissection of QPACK decoder instruction stream
Previously, the HTTP/3 dissector was dissecting only the QPACK encoder
instruction stream, but not the decoder one. The reason for this assymetry
is that only the QPACK encoder stream is necessary for maintaining the state
of the nghttp3 qpack decoder, and dissecting the instructions turned out
to be helpful. On the other hand, the decoder stream isn't needed to maintain
the state of the decoder, hence it has been postponed.

It also simplifies the displayed labels of the QPACK-related protocol items,
opting for the `QPACK ENC[inst count]`/`QPACK DEC[inst count]` terminology.

Finally this MR cleans up the iteration loop over the QPACK instruction stream:

1. The existing code calls the QPACK instructions "opcodes". A QPACK instruction
   has a length and opcode, but is not an opcode by itself.
   This MR makes the names in the code match the terminology used in the RFC.
2. The exisitng code did some unusal calculations for the offsets.
   This MR simplifies the iteration by determining the "sentinel offset",
   and continuing the iteration until either of the following occurs:
   a. The offset of the next instruction would exceed the "sentinel" offset; or
   b. The last instruction in the stream is fragmented; or
   c. Decoding of an instruction fails.