RTSP+SDP: Handle dynamic RTP payload types via "control" media attribute
The "control" media attribute in RTP is IANA registered and uniquely
identifies RTSP. When it is present, do not setup an RTP conversation
(the port is likely 0 anywy), but store the media descriptions under
the URI given for RTSP.

The URI given in the control attribute can be relative; determine
the base URI according to RFC 2326.

When setting up the RTP conversation (in the non-interleaved case)
from RTSP, retrieve the RTP dynamic payload map.

This handles the non-interleaved case of #5208. The interleaved case
(which is more common these days due to firewalls) requires some
additional work, so as to add the RTP dynamic payload maps to the
conversation data for RTP without changing the conversation dissector.