RTSP: Do some request-response tracking
Link the requests and responses, including the request URL.

Get rid of an "if (tree)" statement that would prevent the
the request URL from being parsed. Consolidate a few temporary
proto_item* in the main dissection function. Add the request/response
tree in the main function so that the elements from the tracking
can be added to that subtree.

Ping #5208. (Accessing the request URL in the response is necessary
in order to associate it with a previous SDP.)