Update README.tapping, add reset to tap-rpcprogs
Clarify `doc/README.tapping` in its use of `ui/cli/tap-rpcprogs.c` as
an exemplar of tap listeners.

Adapt `tap-rpcprogs.c` to use an "instance identifier" as described in
the README. It has only ever used a static variable for its state, but
this allows it to actually demonstrate the principle which the README
says it demonstrates.

Add reset and finish routines to `tap-rpcprogs.c` to allow it to
function as expected if tshark is eventually made to reset its taps
(see #20432).