print: Write an empty string for raw JSON hex with null tvb
json_write_field_hex_value gets called to write either the value of an
object, or in some cases to write a value in an array. When writing to
an array for the JSON Raw output, each array index has a specified
meaning - the first index is the value, then the later indices are
the position, length, bitmask, and type.

If called, it *must* write something, otherwise either there will be
an error from the json dumper getting unbalanced in the former case
( json_dumper error: setting name twice on an object member), or
all the subsequent array indices will be shifted over and json2pcap.py
or other consumers will either interpret the indices incorrectly or
possibly throw an exception because the array length is shorter than
expected. (At best, they will ignore the bad data.)

Several of the other potential callers avoid this situation in most
cases by not calling the function if the field_info has length 0, but
-T jsonraw calls the function with zero length if the ordinary values
are not used.

Write an empty string, which is what is currently done if the field
has a length but its data source tvb is not found in the list of data
sources (because the tvb was not added with add_new_data_source.)
It might make sense to use JSON null instead, though.