print: Add the data source index to jsonraw output
The byte offset of a field is relative to its data source. Include
the data source index in the jsonraw information. This is necessary
for tools like json2pcap.py to be able to make changes to a field
(by ignoring all fields with an index other than 0), or for other
purposes.