expert: Ensure more severe ei is displayed in expert column
The original solution to #7733 (commit 586df977d1e616e0768b99f0581)
tries to set the value of COL_EXPERT if the newly added expert info's
severity is greater than (or equal to) its item or that of any of its
descendants. Because items can be faked (and often are when dissecting
the columns), it also always set the column when expert items were
added to faked items.

What we really want to do for the column is, as described, have the
highest severity of any item in the entire frame. Since items can
be faked, but in expert_set_info_vformat we always have the packet info
struct at this point, store the highest expert severity in the packet
info struct and use it.

Fix #20465