Qt: Fix relatedPacketDelegate size hint
If the relatedPacketDelegate is set on a column, then
QTreeView::sizeHintForColumn already calls it for all the
indexes in the row. That means we can't add the result from
relatedPacketDelegate::sizeHint (even for an invalid index)
to the result from QTreeView::sizeHintForColumn.

The right thing is to just have RelatedPacketDelegate::sizeHint
return the correct size.

The correct hint, with some styles, is a bit different than just
adding the decoration size, because some (many?) styles put some
padding between the decoration and the text. Instead of adding
in the known fixed decoration size, properly initialize our
style option and call sizeHint.

Fix #20466

(There's a separate issue mentioned there about how the related
packet symbols are always associated with the first column, not
the first visible column, and thus aren't shown if the first column
is not visible.)