TCP: Handle Microsoft Large Send Offload partial checksums
Microsoft Large Send Offload segmenting TCP packets larger than
the MTU uses a partial checksum calculation that uses zero for
the length field in the pseudo-header, add back in the length after
splitting as well as the payload. This is different to how Linux
handles large TCP packets.

Check for that partial checksum as well.

This could be sped up a bit by reworking the functions.