Qt: Workaround apparent Qt bug in delgate sizeHint with decoration
QStyledItemDelegate::sizeHint requires a nonzero decoration height
(thus, a valid decoration size) in order to properly account for the
decoration when returning the sizeHint. However, a nonzero decoration
height causes the sizeHint height to be increased by at least two
pixels, even if the decoration height plus two pixels is still less
than the sizeHint for the undecorated delegate. This is intentional
by Qt, to "prevent icons from overlapping":
https://github.com/qt/qtbase/blob/a0e0425a107aebf8727673505ea2376400b54b07/src/widgets/styles/qcommonstyle.cpp#L5027

Work around this by calculating the sizeHint twice, once without and
once with the decoration, and taking the width from one and the height
from the other. (We have to add the decoration in the size hint, because
adding it in initStyleOption makes it get used automatically elsewhere
and still increases the height.)

Followup to 25c29cb503cd3a7e3d7de5b70c3aa3ae8a090b7b