dtls, tls: show the entire decrypted plaintext
In DTLS 1.2 with the connection_id extension and (D)TLS 1.3, the plaintext
contains not only the record content, but also a byte containing the content
type and optional padding. Currently, however, Wireshark does not show this
extra data. It does add the content type to the proto tree, but only as a
generated item (despite it not really being generated), making it unclear to the
user where this value comes from.

Correct this with a few changes:

* In `ssl_add_record_info`, store the entire decrypted plaintext, rather than
  just the content. Add a new record field, `plain_data_len`, to store the
  full plaintext length. To avoid ambiguity, rename the old `data_len` field
  to `content_len`.

* In `ssl_get_record_info`, include the entire plaintext in the returned TVB.

* In `dissect_{dtls,dtls13,ssl3}_record`, use a regular item for the content
  type, referencing the aforementioned TVB. Then shrink the TVB
  to cover just the content, so that all subsequent code works as before.