echo,discard: Try Wake On Lan first on UDP
Wake on LAN is very commonly used on UDP ports 7 and 9, which are
registered to the Echo and Discard services. The dissectors for
those services do very little, and never reject a packet, whereas
the WOL dissector is very discriminating and rejects lots of
packets.

Just try Wake on LAN first on those two UDP ports.

Also, register the discard by dissector by name.

Fix #19535