isobus: Use a wmem_strbuf_t instead of snprintf
The second parameter is snprintf is the size of the buffer including
the terminating null; it writes up to bufsz - 1 characters (not
including the terminating zero), unless bufsz is 0. Even though
the minute offset is supposed to be in the range -59 to 59, by the
calculation given it could be -100 or less, thus needing another
character. (In those cases it would be truncated to "-100 mi".)

It would be safe here to use 9 for the second parameter of snprintf,
as the buffer does have size 9, but just use a wmem_strbuf_t instead.

Prevents a format-truncation warning.
Ping #20175