SOME/IP-SD: Avoid format-truncation warnings
The buffers for the numbers as strings don't need to be 128 bytes
wide. Shrinking them keeps gcc from warning about format truncation.

Also, presumably the string should be written even if the event group
name lookup fails. Fix a few cases where the wrong buffer size is
passed to snprintf (a reason not to use snprintf is that it's difficult
to remember the correct parameters).

Ping #20175