snort: Prevent format truncation warning
Instead of having the snort dissector do its own calculation for
adding the fractional part of the absolute time stamp seconds,
call format_fractional_part_nsecs. That function handles format
truncation, and so prevents the gcc warning.

That function includes the decimal point, so adjust for that by
increasing the buffer size by 1 and changing where the buffer is
added.

This is currently the last format truncation warning.
Ping #20175