tls: fix dissection of decrypted handshake messages
`dissect_tls_handshake` uses `is_encrypted_handshake_message` to determine
whether the message is encrypted. When an unfragmented handshake message
(such as Finished) occurs after a Change Cipher Spec message, the logic
in that function assumes that the message is encrypted, so
`dissect_tls_handshake` refrains from dissecting it further.

However, this logic doesn't make sense when the message had already been
decrypted by Wireshark (which is the case when `maybe_encrypted` is false).
In this case, it couldn't possibly still be encrypted, but Wireshark refuses
to dissect it anyway.

Fix this by only calling `is_encrypted_handshake_message` when `maybe_encrypted`
is true. This also makes passing `maybe_encrypted` to
`is_encrypted_handshake_message` pointless, so remove that parameter.

Note that this fix is only relevant for TLS 1.2 and older, since in TLS 1.3
Change Cipher Spec is a no-op.