rdp: various display improvements
This patch reworks the display of clientNetworkData so that the name of the channel
is more obvious. We also show the channelId. It also fixes the info column when
there's multiple messages.
And finally it registers the RDP dissector in heuristic lists of TLS and websocket,
preparing the correct dissecting of ARM traffic (RDP on top of websocket).