nstime: Add functions to round to a certain precision
Add functions to round a nstime_t object to a given precision.
Note the existing functions to print a time truncate the to the
given precision rather than rounding. To round, this function
could be called first.

nstime_t values do not currently carry an indication of the precision
(see the comments in #15579), so this explicitly rounds the nsecs value
as stored. If the nstime_t struct is changed to carry the precision
as a member, this function would presumably change that as well.
(That nstime_t values do not carry an indication of their precision
internally also means that the printing functions are passed some kind
of value obtained from, e.g., a file header.)