PROFINET: add support for displaying PNIO Data without AR
For a full dissection of the PROFINET IO payload in Wireshark, it is
necessary to capture the entire PROFINET connection setup – specifically,
the complete Application Relation (AR) . In practice, capture files often
consist of only a partial recoding of a complete communication cycle,
which means that the payload can’t be shown. This issue is resolved with the
proposed solution.

Instead of relying on the Data from the AR the required information is now
extracted from the Generic Station Description (GSD) using the GMarkup parser.
The GMarkup parser doesn’t create any new external dependencies! The parser
implements two callback functions, one for opening and one for closing
markup elements. When a key element is opened, its attributes are iterated
through and saved into a ParserState object. Upon closing, these values are
transferred into corresponding objects stored in a list within the ParserState.
Later, these objects are used to generate the definitive IOData objects, as
well as IOCS and IOPS objects.

Distinction Between Manual and Heuristic Approaches:

Heuristic Method (applicable only to the Wireshark UI):
For an accurate display of the PNIO payload, it is essential to determine which
MAC address belongs to the controller and which to the device. In the
heuristic approach, this information is extracted from the acyclic frames, as
these frames are always initiated by the controller. In the appropriate
dissector functions, the MAC addresses are saved into a global list. Before
extracting the GSD, it iterates over this global list to verify whether each
packet is an input or an output packet. Based on this, an appropriate
conversation and station_info object are created. All the extracted Data
from the GSD is used to generate the corresponding objects for representing
the PNIO payload, and these objects are stored within the station_info.

Manual Method:
The manual approach is divided into two use cases:

First Scenario:
There are no acyclic frames in the capture or if you using tshark.
In this case, an XML configuration file must be created to manually specify all
conversations, detailing the MAC addresses of both the controller and the
device. For example:
<conversationList>
    <conversation device="XX:XX:XX:XX:XX:XX" controller="XX:XX:XX:XX:XX:XX"/>
    <conversation device="XX:XX:XX:XX:XX:XX" controller="XX:XX:XX:XX:XX:XX"/>
</conversationList>

Second Scenario:
The GSD includes only the AllowedInSlots attribute.
The number of modules for a station can be defined using three attributes.
With the FixedInSlots and UsedInSlots attributes, the number is clearly
described and unambiguous. When using the AllowedInSlots attribute, the number
is not fixed; it only shows a possible slot assignment. Only during the AR the
actual number and position of the modules are communicated. This means that
it is not possible to generate the necessary information from the GSDs with
only AllowedInSlots attributes.

For this purpose, the user should only transfer the information from the GSD to
the Config file that is contained in the PNIO network traffic. The GSD
structure should be retained and modules should then be described using
FixedInSlots, in which they are actually located. The solution now extracts all
the data from this config file instead from the GSD.

It is necessary to create a “current_fake_aruuid” to distinguish the conversation
and station info from a possible AR related conversation.

Cast unused parameters to void to suppress warnings

Cast unused parameters to void to suppress warnings