USB: Fix dissection of USB DFU download request
According to the DFU 1.1 specification, the receipt of
a DFU_DNLOAD request with wLength = 0 implies that the
host is informing the device that there is no more data
to download. However, the USB DFU dissector incorrectly
reports this as a malformed packet.

Reported in #18577.

Signed-off-by: Sahil Siddiq <sahilcdq@proton.me>


(cherry picked from commit 96ea0c38ab47c16a9c910f822cf6e5904245400e)

Co-authored-by: Sahil Siddiq <sahilcdq@proton.me>