Qt: Only auto widen the Conversation/Endpoint columns, don't narrow
The Conversation and Endpoints dialogs auto resize the columns to
fit the data, which includes auto resizing for newly arriving packets
in a live capture.

A user might want to manually widen the columns, especially the
Rel/Abs Start and Duration columns in the Conversation dialog, which
contain a timeline graph. Only automatically widen a column if it's
not narrow enough for the contents, don't shrink them.

Double-clicking a column or selecting "Resize all columns to contents"
from the context-menu still resizes (including shrinking), this is
only for the automatic resizing when new packets arrive.

Automatically resize the columns to match the header size when
first creating the tree.

Fix #15978


(cherry picked from commit 4d7b01df49a3ab2b13658debe70139e2346f0f39)

Co-authored-by: John Thacker <johnthacker@gmail.com>