tap-camelsrt: improve memory management
Expose the existing function `camelsrt_reset()` as the tap reset
function. Add a `camelsrt_finish()` function for the tap finish
function.

Remove `filter` from the tap-state structure since it wasn't used for
anything. (And it was leaked if `camelsrt_reset()` was called, since it
naively zeroes out the entire structure, including the pointer to the
allocated filter string.) `register_tap_listener()` already takes a copy
of the filter string so we don't need to allocate and track another
copy.

Replace `0` with `TL_REQUIRES_NOTHING` in the call to
`register_tap_listener()` to improve code self-documentation.