tap-credentials: fix memory management
Remove a global `wmem_array` holding credentials. Create a listener
tap-data structure to hold that array instead on a per-listener basis.

The existing `credentials_reset()` function did nothing, since it was
passed the tap *listener's* tap-data structure, which was set to `NULL`
by the call to `register_tap_listener()`. It appears to have expected to
receive the actual *tap's* data, being a `tap_credential_t *`, so
refactor its contents into a new function `tap_credential_free()`.
Call this revised function on each element of the credentials array
when resetting or finishing the tap listener.

Adapt the call to `register_tap_listener()` so it creates and passes in
a listener tap-data structure, and also specifies a finish function (the
new `credentials_finish()`).

`credentials_packet()` allocates memory for a clone of the credential
statistics coming from the tap, and calls `wmem_array_append()` to store
it. But `wmem_array_append()` fills the new array element with a copy of
the passed-in item, so the clone was being leaked. Add a comment
explaining this, and free the clone.

Remove editor modelines that incorrectly specified the file as being
tab-indented instead of space-indented.

Ping #20432