Dissector tables: Show the integer table entries in the configured base
At dissector table creation a base is configured in which the entries
are to be shown in the UI and entered in the Decode As dialog. However,
when showing or printing the tables, this configuration is not applied,
which may lead to difficulty finding/recognizing entries.

Extend the table creation and printing functions, for integer indexed
tables, to take the configured base, and type, into consideration when
showing values.

In this context packet.c:dissector_dump_dissector_tables_display()
contained code to handle BASE_HEX_DEC and BASE_DEC_HEX basis. Even
though valid, these are not used nor supported anywhere, so can be
dropped as well. Support for BASE_OCT was missing, so that is added.