MACsec: Restore behavior of authentication-only frames
Add the ethertype in the unencrypted, auth-only frame even if we
don't have the key (as we dissect using it anyway), as was done
before decryption was added.

Add the MACsec Ethertype to the macsec tree instead of the root tree.

Add the ICV item and verification item to the macsec tree after the
Ethertype, to preserve ordering.

Only call the data dissector for undissected payload data, as was
done before decryption was added. (If a separate item for "encrypted
data" is requested, that could be added later as a new hf. It makes
more sense to register that with macsec instead of using the generic
"data" field.)

Don't try to authenticate or dissect if the datalink type isn't
AT_ETHER. (That requires some odd configuration, but avoid a possible
crash.) In those cases, also don't memcpy the IV, ICV, payload, etc.
out of the TVB.

Eliminate a redundant memcpy to a header.

Cisco, and possibly others, support MACsec on other EtherTypes, to
avoid conflicting with provider bridges. Check pinfo->match_uint
to see if the dissector was called for a different EtherType.

MACsec now supports on Jumbo frames. The max payload length should
probably be increased, but at least check to make sure we're not
going to run off the end. (These should probably be allocated with
pinfo->pool memory rather than big static arrays, as the fast block
allocator is fast.)

Add some comments