Qt: Allow adjusting visual sample rate in RTP player
Add a spinbox, defaulting at the current value of 1 kHz,
to control the sample rate used for the waveform display.
This is useful if the most interesting harmonics in a file,
e.g. a test tone, are above the Nyquist frequency for the
the default sample rate.

It might make sense to store this and some other controls in
RTP player that cause recalculation (playback timing, output
audio rate) in recent.

Move a few things that always need to be done at the beginning
and end of the visual resampling into decodeVisual.

Fix #20502