epan: Allow field_info to contain full 64 bit width offset
Part of the flags field in the field_info structure is being used to
record how many bits the field actually covers and what its offset
from the start of the field is. However, while there are enough bits
allocated for a 64 bit width field, the offset only allows for eight.
Therefore only an offset within a byte can be given. The start of the
field is not adjusted if the offset overflows a byte.

Keeping all things equal, and since there is enough room left in this
flags field, extend the offset part of the flags field to accommodate
64 bit offsets.

Fixes #20507