dfilter: Don't use an incorrect token for xor and not in nodes
Fix translation of filters containing XOR with the dfilter
translator.

Operators often have two or more possible tokens that have the
same semantics, one symbolic and one a word. Syntax tree nodes
for operators and tests save the original token and use it when
transforming the node to a string in certain circumstances instead
of just one canonical string, so that, e.g., error messages are
more consistent with what was typed in.

(Note that the Boolean tests are never used in error messages
from the semantic check because any valid expression has a Boolean
value and can be used in a test so long as the basic syntax is
correct.)

This string output is also used in the dfilter translator.

The XOR test is transformed into a set of AND, OR, and NOT nodes.
Don't copy the XOR token as the token for the central AND node,
or else the dfilter translator will output the original token
for the AND node, resulting in an incorrect expression.

Similarly, the combination of the "not" and "in" tokens are combined
into a single node. Merely copying the "in" node results in the
string representation for the node lacking the "not".
We don't need to worry about saving the location, as these are
infix operators and the location is derived from merging with
the surrounding nodes.

We could perhaps check the xor and not nodes in the two cases so
as to consistently use symbols or words for tokens for the derived
nodes.