FT_ABSOLUTE_TIME: support special time values with special meanings.
Allow a time_value_string table to be associated with FT_ABSOLUTE_TIME
fields; if the time has one of the values in the table, display it as
the corresponding string, rather than as the specified date and time.

This can handle, for example, certain Zigbee fields, where a value of 0
(corresponding to the Zigbee epoch, 2000-01-01 00:00:00 UTC) means
"Now", and, for some of those fields, a value of 0xFFFFFFFF means
"cancel a previous request".