Zigbee: use FT_ABSOLUTE_TIME for Zigbee UTCTime values.
Use the new "time value strings" mechanism to handle Zigbee time values
0x00000000 and 0xFFFFFFFF specially as appropriate.

In the ImageBlockResponse, handle Request Time as a UTCTime if the
Current Time field is non-zero.

Use ENC_TIME_ZBEE_ZCL to fetch a Zigbee time stamp in the GBCS
dissector.

Add TIME_VALS to the checkAPI script (and fix the error message to
mention VALS_EXT_PTR).