ICMP, ICMPv6: Add expert info for error types
Add an expert info for error types, which are also indicated in the
default coloring rule. This creates an entry in the Expert Info
Severity column and the Expert Info dialog.

Fix #20480

(Yes, there might be more severe expert info severities in the frame
not shown in the Expert Info column, but that's a consequence of all
the columns being set non-writable before dissecting encapsulated
error packets.)