pcapng: Added dissection of Darwin metadata
On Darwin-based systems, running `tcpdump -k` will collect useful
metadata, including:
 - Process information (pid, name, etc.)
 - Packet metadata (Wake, Retry, First packet in flow, etc.)
 - The id of the flow the packet belongs to
 - Trace tag, which can be set by the kernel

 In addition, if `tcpdump` is using the `droptap` pseudo-interface
 for capturing packet that were dropped by the kernel, the metadata will
 include:
 - The drop reason code (for now, opaque)
 - The function name that dropped the packet (not always present)
 - The line number of the callsite of the packet drop (likewise, not always
   present)

These abilities were added to Darwin `tcpdump` about 13 years ago, and
at the time the pcapng format did not support custom blocks that can be
attributed to companies via the Private Enterprise Number (PEN).

Because of that, Darwin `tcpdump` uses the "local" codes for the block
that stores the process information, as well as for the EPB options that
describe the metadata, or refer to the previously mentioned blocks.

This presents a problem, since the "local" codes can be used by
3rd-party plugins.

In order to disambiguate the meaning of the "local" codes, the plugins
are given priority in processing the pcapng block data. If no plugin
claims the "local" codes, the block data is assumed to belong to Darwin
systems.

In the future the disambiguation mechanism can be enhanced, e.g. via
preferences, but at the moment it is kept simple.

Currently, the MR supports reading and writing DPEB information, but not
merging.